build: Add documentation to kbuild.pl script
This tool is an enigma without documentation rendering it useless to
others without a good amount of analysis.  This patch adds -h and --help
usage help.  However, being that I have only recently studied it I would
like for somebody else who is familiar with the tool to kindly review it
for correctness.